refactor
Refactored Binance coins into seperate methods into BinanceAPI
