+++ crypto/src/main/java/com/crypto/crypto/web/BinanceAPI.java
@@ -5,6 +5,12 @@ import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+
+import com.crypto.crypto.dto.BinanceCoinDataDTO;
+import com.crypto.crypto.service.BinanceService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,6 +19,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class BinanceAPI {
+  
+  private final BinanceService binanceService;
 
   public String getData(String coin, String date) throws IOException, InterruptedException {
     //호출할 코인 URL 빌드
@@ -40,5 +48,69 @@ public class BinanceAPI {
 
     System.out.println("Binance " + coin + " 요청");
     return response.body();
+  }
+  
+  public void buildHistory(String[] coins) throws  Exception{
+    LocalDateTime localDateTime = LocalDateTime.parse("2023-02-10T00:00:00");
+    String date = String.valueOf(localDateTime.atZone(ZoneId.of("UTC")).toInstant().toEpochMilli());
+  
+    for(int i = 0; i< coins.length; i++){
+      while(true){
+        String data = getData(coins[i], date);
+    
+    
+        data = data.replace("[[", "[");
+        data = data.replace("]]", "]");
+    
+        String[] coinData = data.split("],");
+    
+        if(date.equals(coinData[0].split(",")[0].replace("[", "")) || data.length() < 3 ){
+          System.out.println("done");
+          break;
+        }
+    
+        date = coinData[0].split(",")[0].replace("[", "");
+    
+        for (int j = 0; j < coinData.length; j++) {
+          coinData[j] = coinData[j].replace("[", "");
+          coinData[j] = coinData[j].replace("]", "");
+          coinData[j] = coinData[j].replace("\"", "");
+          String[] dailyData = coinData[j].split(",");
+          BinanceCoinDataDTO binanceCoinDataDTO = binanceCoinDataDtoFromStringArray(dailyData, coins[i]);
+      
+          binanceService.addData(binanceCoinDataDTO);
+        }
+      }
+    }
+  }
+  
+  public LocalDateTime milliToLocalDateTime(String milli) {
+    String result = String.format("%.0f", Double.parseDouble(milli));
+    
+    return Instant.ofEpochMilli(Long.parseLong(result))
+            .atZone(ZoneId.of("UTC"))
+            .toLocalDateTime();
+  }
+  
+  public BinanceCoinDataDTO binanceCoinDataDtoFromStringArray(String[] dailyData, String coinName) {
+    LocalDateTime openTime = milliToLocalDateTime(dailyData[0]);
+    LocalDateTime closeTime = milliToLocalDateTime(dailyData[6]);
+    
+    BinanceCoinDataDTO binanceCoinDataDTO = BinanceCoinDataDTO.builder()
+            .coin(coinName)
+            .openTime(openTime)
+            .openingPrice(dailyData[1])
+            .highPrice(dailyData[2])
+            .lowPrice(dailyData[3])
+            .closePrice(dailyData[4])
+            .volume(dailyData[5])
+            .closeTime(closeTime)
+            .quoteAssetVolume(dailyData[7])
+            .numTrades(dailyData[8])
+            .takerBuyBaseAssetVolume(dailyData[9])
+            .takerBuyQuoteAssetVolume(dailyData[10])
+            .build();
+    
+    return binanceCoinDataDTO;
   }
 }
